pyproject.toml
[tool.poetry]
name = "jampdf"
version = "0.1.0"
description = "A simple static site generator producing PDF files ready to be shared in an organized fashion."
authors = ["Sebastian Grötzinger <lowfatprophet@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
weasyprint = "^63.0"
markdown = "^3.7"
click = "^8.1.7"
jinja2 = "^3.1.4"
python-frontmatter = "^1.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.poetry.scripts]
jampdf = "jampdf.cli:cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"